requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

# ORM / DB
SQLAlchemy==2.0.29
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
itsdangerous>=2.1.2

# Async Database Drivers
aiosqlite==0.20.0   # local development only
pyodbc
aioodbc

# Azure SDKs
azure-storage-blob==12.19.1
azure-core==1.30.1
azure-identity==1.17.1

# Security & Password Hashing
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Utilities
requests==2.32.3
python-multipart==0.0.9
aiofiles==23.2.1
Jinja2==3.1.3

# REMOVE (PostgreSQL — not needed)
# asyncpg==0.29.0

# REMOVE (Not compatible with Azure SQL)
# pytds
# python-tds

# OPTIONAL — keep only if actually used
# python-jose==3.3.0
# PyJWT==2.8.0
